moved auto reducer code in its own package
